Realtime instead of cycles in guardian attack

+++ TowerDefense/TowerDefense.Source/Guardians/States/LoadingState.cs
@@ -8,7 +8,7 @@ namespace TowerDefense.Source.Guardians.States
 {
     public class LoadingState : State
     {
-        public LoadingState(int initialCounter) : base(initialCounter) { }
+        public LoadingState(long initialCounter) : base(initialCounter) { }
 
         public override List<Projectile> InitiateAttack(Vector2 target, int targetSpeed) => new List<Projectile>();
